Add back missing popover-message style and convert to RLT
