+++ packages/components/src/internal/components/samples/DisableableMenuItem.tsx
@@ -45,7 +45,7 @@ export const DisableableMenuItem: FC<DisableableMenuItemProps> = memo(props => {
     const overlay = (
         <Popover
             id="disable-operation-warning"
-            className="disabled-menu-item-popover"
+            className="disabled-menu-item-popover popover-message"
             placement={placement}
             targetRef={targetRef}
         >
